requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23

# AI/ML Libraries
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Image Processing
pytesseract==0.3.10

# Cloud Download
gdown==4.7.1
requests==2.31.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
